+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html). However, all releases before 1.0.0 have breaking changes
 between minor-version updates.
 
+## [2.2.0] - Unreleased
+
+### Changed
+
+- Changed location preference rules to use probability scores and arithmetic mean calculations
+- Changed life event considerations to use arithmetic mean calculations
+
 ## [2.1.1] - 2023-30-10
 
 ### Fixed